# Conflicts:
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/availableproductspage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/categorypage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/homepage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/listOrdersPage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/newOrderPage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/newsItemPage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/newsPage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/outletpage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/productpage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/searchproductspage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/statusOrderPage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/categoriespage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/categorypage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/clientgrouppage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/clientgroupspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/clientpage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/clientspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/editorderpage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/inventoriespage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/inventorypage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/newscategoriespage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/newscategorypage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/newsitempage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/newspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/orderpage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/orderspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/outletpage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/outletspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/productattributeitempage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/productattributepage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/productattributespage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/productpage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/productspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/settingspage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/warehousepage.js
#	be/src/Project/Web/Seller/Seller.Web/wwwroot/dist/js/warehousespage.js
#	fe/src/project/AspNetCore/areas/Orders/components/StatusOrder/StatusOrder.js

+++ be/src/Project/Services/Inventory/Inventory.Api/ServicesModels/InventoryServiceModels/InventoryServiceModel.cs
@@ -11,9 +11,9 @@ namespace Inventory.Api.ServicesModels.InventoryServiceModels
         public string Sku { get; set; }
         public Guid? WarehouseId { get; set; }
         public string WarehouseName { get; set; }
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
         public string Ean { get; set; }
-        public int? AvailableQuantity { get; set; }
+        public double? AvailableQuantity { get; set; }
         public int? RestockableInDays { get; set; }
         public DateTime? ExpectedDelivery { get; set; }
         public DateTime LastModifiedDate { get; set; }
