+++ be/src/Project/Services/Inventory/Inventory.Api/ServicesModels/OutletServiceModels/UpdateOutletServiceModel.cs
@@ -10,10 +10,10 @@ namespace Inventory.Api.ServicesModels.OutletServiceModels
         public Guid? ProductId { get; set; }
         public string ProductName { get; set; }
         public string ProductSku { get; set; }
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public int AvailableQuantity { get; set; }
+        public double AvailableQuantity { get; set; }
         public string ProductEan { get; set; }
     }
 }
