+++ be/src/Project/Web/Seller/Seller.Web/Areas/Inventory/DomainModels/OutletItem.cs
@@ -11,11 +11,11 @@ namespace Seller.Web.Areas.Inventory.DomainModels
         public string ProductName { get; set; }
         public string ProductSku { get; set; }
         public Guid SellerId { get; set; }
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
         public string Ean { get; set; }
-        public int AvailableQuantity { get; set; }
+        public double AvailableQuantity { get; set; }
         public DateTime LastModifiedDate { get; set; }
         public DateTime CreatedDate { get; set; }
     }
